+++ src/api/comment.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 
-export const FetchAllCommentsData = async ({id,page=1}) => {
+export const FetchAllCommentsData = async ({id}) => {
     try {
         const user = JSON.parse(localStorage.getItem("user"));
         const accessToken = user?.access_token || "";
@@ -8,18 +8,15 @@ export const FetchAllCommentsData = async ({id,page=1}) => {
             console.error("Access token not available");
             return null;
         }
-     
-        const res = await axios.get(`http://localhost:8000/comments/?post_id=${id}&page=${page}`, {
+
+        const res = await axios.get(`http://localhost:8000/comments/${id}/:`, {
             headers: {
                 Authorization: `Bearer ${accessToken}`,
             },
         });
 
-             
-console.log("res.dataFetchAllCommentsData")       
-console.log("res.dataFetchAllCommentsData")       
-console.log("res.dataFetchAllCommentsData") 
-console.log(res.data)
+       
+
         // rest of the code...
         return res.data;
     } catch (error) {
@@ -30,7 +27,7 @@ console.log(res.data)
 
 
 
-export const FetchAllReqCommentsData = async ({id,page=1}) => {
+export const FetchCreateComments = async ({ post_id, recurit_id, page ,content,to_user}) => {
     try {
         const user = JSON.parse(localStorage.getItem("user"));
         const accessToken = user?.access_token || "";
@@ -38,68 +35,26 @@ export const FetchAllReqCommentsData = async ({id,page=1}) => {
             console.error("Access token not available");
             return null;
         }
-        console.log("id,page")
-        console.log("id,page")
-        console.log("id,page")
-        console.log("id,page")
-        console.log("id,page")
-        console.log("id,page")
-        console.log("id,page")
-        console.log("id,page")
-     console.log(id,page)
-     console.log("id,page")
-     console.log("id,page")
-     console.log("id,page")
-     console.log("id,page")
-        const res = await axios.get(`http://localhost:8000/comments/?recruits=${id}&page=${page}`, {
+   
+
+        const res = await axios.post("http://localhost:8000/comments/", {
+            "post_id":{post_id},
+            "recurit_id":{recurit_id},
+            "page":{page},
+            "content":{content},
+            "to_user": {to_user}
+        }, {
             headers: {
-                Authorization: `Bearer ${accessToken}`,
-            },
+                Authorization: `Bearer ${accessToken}`
+            }
         });
 
-             
-console.log("res.req")       
-console.log("res.req")       
-console.log("res.req") 
-console.log(res.data)
+  
         // rest of the code...
         return res.data;
     } catch (error) {
-        console.error("Fetching notice failed:", error.message);
+        console.error("Fetching Create failed:", error.message);
         return null;
     }
 };
 
-
-
-export const FetchCreateComments = async ({ post_id, recruits, content, to_user }) => {
-    try {
-      const user = JSON.parse(localStorage.getItem("user"));
-      const accessToken = user?.access_token || "";
-      if (!accessToken) {
-        console.error("Access token not available");
-        return null;
-      }
-  
-      // post_id를 숫자로 변환
-      const numericPostId = Number(post_id);
-      const numericRecuritId= Number(recruits);
-      const res = await axios.post("http://localhost:8000/comments/", {
-        "post_id": numericPostId,
-        "recruits": numericRecuritId,  // recruits가 이미 숫자라면 변환 필요 없음
-        "content": content,
-        "to_user": to_user
-      }, {
-        headers: {
-          Authorization: `Bearer ${accessToken}`
-        }
-      });
-  
-      // rest of the code...
-      return res.data;
-    } catch (error) {
-      console.error("Fetching Create failed:", error.message);
-      return null;
-    }
-  };
-
