+++ src/components/CommentList.jsx
@@ -10,7 +10,7 @@ const CommentList = ({ comments }) => {
           <div key={index} className='flex items-start mb-2'>
             <div className='bg-black w-8 h-8 rounded-full mr-2'></div>
             <div>
-              <strong className='font-semibold'>{comment.user_id}</strong>
+              <strong className='font-semibold'>{comment.user}</strong>
               <p className='font-mono'>{comment.content}</p>
             </div>
           </div>
