+++ src/pages/RecruitmentDetailPage.jsx
@@ -7,64 +7,17 @@ import Contour from '../components/ui/Contour.jsx';
 import { Link, useParams } from 'react-router-dom';
 import { FetchPostData } from '../api/post.js';
 import { FetchRecruits } from '../api/recruits.js';
-import { FetchAllReqCommentsData, FetchCreateComments } from '../api/comment.js';
 import { FetchTeam } from '../api/team.js';
-import { submitReport } from '../api/report.js';
-import { likePost, unlikePost } from '../api/likes.js';
-import { checkRecruitApplication } from '../api/applyRecruit';
-import { applyForRecruit } from '../api/applyRecruit';
 
 function RecruitmentDetailPage() {
-
   const { screenSize } = useWindowSize();
   const { id } = useParams();
   const [data, setData] = useState(null);
   const [teamData, setTeamData] = useState(null);
-  const [isLiked, setIsLiked] = useState(false);
   const [comments, setComments] = useState([
     { user: '@사용자ㄷㄷㄷㄷ', content: 'ㄴㄴㄴㄴ 님 안녕하세요' },
     // Add other comments as needed
   ]);
-  const [currentPage, setCurrentPage] = useState(1);
-  
-  const [inserComment, setInserComment] = useState('');
-  const [AmIRecruit, setAmIRecruit] = useState('');
-  const [AmIView, setAmIView] = useState(1);
-
-  const handleInputChange = (event) => {
-    setInserComment(event.target.value);
-  };
-  const postId = id
-
-
-  const RecruitApplication = async () => {
-    try {
-      const IRecruit = await checkRecruitApplication(id );
-      console.log("IRecruit")
-      console.log("IRecruit")
-      console.log("IRecruit")
-      console.log("IRecruit")
-      console.log("IRecruit")
-      console.log("IRecruit")
-      console.log(IRecruit)
-      console.log("IRecruit")
-      console.log("IRecruit")
-      console.log("IRecruit")
-      console.log("IRecruit")
-      console.log("IRecruit")
-      console.log("IRecruit")
-      setAmIRecruit((prev) => ({
-
-        IRecruit,
-      }));
-
-
-    } catch (error) {
-      console.error('Error fetching post data:', error);
-    }
-  };
-
-
   useEffect(() => {
     const getPost = async () => {
       try {
@@ -77,9 +30,6 @@ function RecruitmentDetailPage() {
         console.error('Error fetching post data:', error);
       }
     };
-
-   
-
   
     const getTeam = async () => {
       try {
@@ -92,83 +42,13 @@ function RecruitmentDetailPage() {
         console.error('Error fetching team data:', error);
       }
     };
-    commentFetch()
+  
     getPost();
     getTeam();
-    RecruitApplication()
-  }, [currentPage]);
+  }, []);
   
-  const AmIReClikc=()=>{
-    RecruitApplication()
-    let a = AmIRecruit.IRecruit==undefined ?AmIRecruit.IRecruit:"신청 상태가 없는"
-    alert( a+"상태 입니다")
-    console.log("AmIRecruit")
-    console.log(AmIRecruit.IRecruit)
-    console.log("AmIRecruit")
-    console.log(AmIRecruit)
-    
-  }
-
- const RecruitBtn=async()=>{
-    applyForRecruit(id)
-  }
- 
-
-
-  const handlePageClick = (page) => {
-    setCurrentPage(page+1);
-  };
-
-  const commentFetch =async()=>{
-    let a = await FetchAllReqCommentsData({id:id,page:currentPage}); // 하드코드
-    setComments({a})
-   
-  }
-
-
-
-  const handleLikeClick = async () => {
-    try {
-      await likePost(postId);
-      setIsLiked(true);
-    } catch (error) {
-      console.error('Error liking the post:', error);
-    }
-  };
-
-  const handleUnlikeClick = async () => {
-    try {
-      await unlikePost(postId);
-      setIsLiked(false);
-    } catch (error) {
-      console.error('Error unliking the post:', error);
-    }
-
-  };
-  const ReportClick = async()=>{
-    console.log("data")
-    console.log("data")
+ console.log(data)
     console.log(data)
-    let a=await submitReport({user_id:data.author,content:`모집글${id}를 신고 당했습니다.`})
-    console.log(a)
-  }
-
-
-  const CreatComment = async () => {
-    console.log("asd")
-    try {
-      await FetchCreateComments({
-        recruits: id,
-        content: inserComment,
-      });
-      await commentFetch(); // Fetch and update comments after creating a new comment
-    } catch (error) {
-      console.error('Error creating comment:', error);
-    }
-  };
-
-
-
   if (!data) {
     // You can render a loading spinner or message here
     return <div>Loading...</div>;
@@ -243,31 +123,16 @@ function RecruitmentDetailPage() {
             
             <Margin top="2" />
             <div className='flex justify-center'>
-            {isLiked ? (
-                   <DynamicColorButton
-                   color="blue"
-                   text="좋아요"
-                   btnstyle="py-2 px-4"
-                   onClick={handleUnlikeClick}
-                 />
-      
-      ) : (
-        <DynamicColorButton
-        className={`py-2 px-4 text-sm`} // Apply text-sm class for smaller text
-        onClick={handleLikeClick}
-        text="좋아요 취소"
-        color="red"
-      >
-      
-      </DynamicColorButton>
-      )}
-
+              <DynamicColorButton
+                color="blue"
+                text="좋아요"
+                btnstyle="py-2 px-4"
+              />
               <Margin left="4" />
               <DynamicColorButton
                 color="red"
                 text="신고"
                 btnstyle="py-2 px-4 mr-2"
-                onClick={ReportClick}
               />
             </div>
 
@@ -277,10 +142,13 @@ function RecruitmentDetailPage() {
                 <DynamicColorButton
                   text="신청 현황"
                   btnstyle="py-1 px-2 flex-shrink-0"
-                  onClick={AmIReClikc}
                 />
                 <Margin left="1" />
-           
+                <DynamicColorButton
+                  color="blue"
+                  text="수정하기"
+                  btnstyle="py-1 px-2 flex-shrink-0"
+                />
                 <Margin left="1" />
                 <DynamicColorButton
                   color="red"
@@ -294,7 +162,6 @@ function RecruitmentDetailPage() {
                   color="blue"
                   text="신청하기"
                   btnstyle="py-1 px-2 ml-0 items-end flex-shrink-0"
-                  onClick={RecruitBtn}
                 />
                 <Margin left="1" />
                 <Link to="/recruits">
@@ -311,35 +178,22 @@ function RecruitmentDetailPage() {
               <div className='text-2xl font-bold mb-4'>댓글</div>
               
               <div className='flex items-center justify-center'>
-              <input
-        className='border p-2 w-3/4 rounded-md'
-        placeholder='댓글 입력...'
-        value={inserComment}
-        onChange={handleInputChange}
-      />
+                <input
+                  className='border p-2 w-3/4 rounded-md'
+                  placeholder='댓글 입력...'
+                />
                 <DynamicColorButton
                   color="black"
                   text="댓글 달기"
                   btnstyle="py-2 px-2 ml-2"
-                  onClick={CreatComment}
                 />
               </div>
             </div>
             
             <div className='flex justify-between items-center'>
-            <CommentList comments={comments.a?comments.a.results:[]} />
+              <CommentList comments={comments} />
             </div>
           </div>
-
-          <div className='flex w-full justify-center items-center'>
-              {Array.from({ length: Math.ceil(comments.a ? comments.a.count / 5 : 1) }, (_, index) => (
-    <span
-      key={index}
-      className={`cursor-pointer mx-1  text-center ${currentPage === index + 1 ? 'font-bold' : ''}`}
-      onClick={() => handlePageClick(index )}
-    >{index+1}</span>
-              ))}
-              </div>
         </div>
       </div>
     </div>
