+++ src/pages/RecruitmentEntryPage.jsx
@@ -32,7 +32,6 @@ function RecruitmentPostEntryPage() {
     setFormData({ ...formData, [field]: e.target.value });
     console.log(formData);
   };
-  
 
 
   const regions = [
@@ -184,7 +183,6 @@ function RecruitmentPostEntryPage() {
               <DynamicColorButton color="blue" text="작성" btnstyle="" onClick={handleSave} />
             </div>
           </div>
-        
         </div>
       </div>
     </div>
