+++ src/pages/RecruitmentListPage.jsx
@@ -36,9 +36,6 @@ function RecruitmentListPage() {
 
     
   });
-  const [Count, setCount] = useState(1);
-
-  
   const handleRegionSelect = (selectedRegion) => {
     setFilters({ ...filters, region: selectedRegion });
     console.log('Selected Region:', selectedRegion);
@@ -85,7 +82,6 @@ function RecruitmentListPage() {
           console.log("fetchDataindata")
           console.log(data)
           setTeamData(data.results);
-          setCount(data.count);
           // setTotalPages(data.total_pages);
         }
         // if(RecruitsData){
@@ -96,9 +92,7 @@ function RecruitmentListPage() {
         console.error('Error fetching team data:', error.message);
       }
     };
-    console.log("Count")
-    console.log("Count")
-console.log(Count)
+
     fetchData();
   }, [filters, currentPage]);
 
@@ -161,44 +155,45 @@ console.log(Count)
             </div>
             <Margin top="1" plustailwind="h-1" />
             <Contour></Contour>
-            <div className='flex  justify-around'>
+            <div className='flex  w-32'>
             <SelectButton
                 btnTitle="region"
                 btnoptions={[
                   '서울',
                   '경기',
                   '강원',
-                  '충남',
-                  '충북', 
-                  '전남',
-                  '전북',
-                  '경남',
-                  '제주',
-            
+                  '충청북도',
+                  '충청남도',
+                  '전라북도',
+                  '전라남도',
+                  '경상북도',
+                  '경상남도',
+                  '제주특별자치도'
                 ]}
                 title=" "
-                size="32"
+                size="15"
                 onOptionSelect={handleRegionSelect}
               />
               <SelectButton
                 btnTitle="Order By"
                 btnoptions={[
-                  'views',
-                  'create_at',
-                
+                  '조회수',
+                  '좋아요',
+                  '마감일',
+                  '작성일',
                 ]}
-                size="32"
+                size="15"
                 title=" "
                 onOptionSelect={handleOrderSelect}
               />
               <SelectButton
                 btnTitle="Sort"
                 btnoptions={[
-                  'desc',
-                  'asc',
+                  '오름차',
+                  '내림차',
                 ]}
                 title=" "
-                size="32"
+                size="15"
                 onOptionSelect={handleSortSelect}
               />
             </div>
@@ -228,21 +223,16 @@ console.log(Count)
       
             </div>
             <div className='flex justify-between items-center'>
-              <div className=' w-full '>
-             
-
-              <div className="flex justify-center items-center w-full text-center">
-  {Array.from({ length: Math.ceil(Count ? Count / 5 : 1) }, (_, index) => (
-    <span
-      key={index}
-      className={`cursor-pointer mx-1  text-center ${currentPage === index + 1 ? 'font-bold' : ''}`}
-      onClick={() => handlePageClick(index + 1)}
-    >
-      {index + 1}
-    </span>
-  ))}
-</div>
-
+              <div>
+              {Array.from({ length: Math.min( 4) }, (_, index) => (
+  <span
+    key={index}
+    className={`cursor-pointer mx-1 ${currentPage === index + 1 ? 'font-bold' : ''}`}
+    onClick={() => handlePageClick(index + 1)}
+  >
+    {index + 1}
+  </span>
+))}
 
               </div>
             </div>
